Добавлено извлечения вырезанных имплантов
Чтобы извлечь имплант, вырезайте его с путсым имплант кейсом во второй руке.

+++ code/modules/surgery/implant_removal.dm
@@ -26,12 +26,25 @@
 /datum/surgery_step/extract_implant/success(mob/user, mob/living/carbon/target, target_zone, obj/item/tool, datum/surgery/surgery)
 	if(I)
 		user.visible_message("[user] successfully removes [I] from [target]'s [target_zone]!", "<span class='notice'>You successfully remove [I] from [target]'s [target_zone].</span>")
-		if(istype(I, /obj/item/weapon/implant/loyalty))
-			target << "<span class='notice'><b>You feel a sense of liberation as Nanotrasen's grip on your mind fades away.</b></span>"
-		qdel(I)
-		if(istype(target, /mob/living/carbon/human))
-			var/mob/living/carbon/human/H = target
-			H.sec_hud_set_implants()
+		I.removed(target)
+
+		var/obj/item/weapon/implantcase/case
+
+		if(istype(user.get_item_by_slot(slot_l_hand), /obj/item/weapon/implantcase))
+			case = user.get_item_by_slot(slot_l_hand)
+		else if(istype(user.get_item_by_slot(slot_r_hand), /obj/item/weapon/implantcase))
+			case = user.get_item_by_slot(slot_r_hand)
+		else
+			case = locate(/obj/item/weapon/implantcase) in get_turf(target)
+
+		if(case && !case.imp)
+			case.imp = I
+			I.loc = case
+			case.update_icon()
+			user.visible_message("[user] places [I] into [case]!", "<span class='notice'>You place [I] into [case].</span>")
+		else
+			qdel(I)
+
 	else
 		user << "<span class='warning'>You can't find anything in [target]'s [target_zone]!</span>"
 	return 1
